+++ support-frontend/assets/components/amazonPayForm/amazonPayFormContainer.tsx
@@ -2,6 +2,8 @@ import { useAmazonPayObjects } from 'helpers/customHooks/useAmazonPayObjects';
 import { AmazonPay } from 'helpers/forms/paymentMethods';
 import { getContributionType } from 'helpers/redux/checkout/product/selectors/productType';
 import { useContributionsSelector } from 'helpers/redux/storeHooks';
+import { trackComponentClick } from 'helpers/tracking/behaviour';
+import { logException } from 'helpers/utilities/logger';
 import AmazonPayForm from './amazonPayForm';
 
 export function AmazonPayFormContainer(): JSX.Element {
@@ -19,33 +21,40 @@ export function AmazonPayFormContainer(): JSX.Element {
 		(state) => state.page.user.isTestUser,
 	);
 
-	console.log(
-		'amazonPayFormContainer.paymentMethod,countryGroupId,isTestUser',
-		paymentMethod,
-		countryGroupId,
-		false,
-	);
 	const { loginObject, paymentsObject } = useAmazonPayObjects(
 		paymentMethod === AmazonPay,
 		countryGroupId,
-		false,
+		userInNewProductTest ? userInNewProductTest : false,
 	);
-	console.log('amazonPayFormContainer.loginObject', loginObject);
+
 	const hasAccessToken = useContributionsSelector(
 		(state) => state.page.checkoutForm.payment.amazonPay.hasAccessToken,
 	);
 
+	if (hasAccessToken) {
+		trackComponentClick('amazon-pay-login-click');
+		const loginOptions = {
+			scope: 'profile postal_code payments:widget payments:shipping_address',
+			popup: true,
+		};
+		if (loginObject) {
+			console.log('amazonPayFormContainer.loginObject.authorize2');
+			loginObject.authorize(loginOptions, (response) => {
+				if (response.error) {
+					logException(`Error from Amazon login: ${response.error}`);
+				}
+			});
+		}
+	}
+
 	return hasAccessToken ? (
-		<>
-			<div>ShowAmazonPayCard</div>
-			<AmazonPayForm
-				amazonLoginObject={loginObject}
-				amazonPaymentsObject={paymentsObject}
-				isTestUser={userInNewProductTest ? userInNewProductTest : false}
-				contributionType={contributionType}
-			/>
-		</>
+		<AmazonPayForm
+			amazonLoginObject={loginObject}
+			amazonPaymentsObject={paymentsObject}
+			isTestUser={userInNewProductTest ? userInNewProductTest : false}
+			contributionType={contributionType}
+		/>
 	) : (
-		<div>HideAmazonPayCard</div>
+		<div></div>
 	);
 }
