+++ support-frontend/assets/components/amazonPayForm/amazonPaymentButton.tsx
@@ -1,35 +1,13 @@
 import { DefaultPaymentButtonContainer } from 'components/paymentButton/defaultPaymentButtonContainer';
-import AnimatedDots from 'components/spinners/animatedDots';
-import { useAmazonPayObjects } from 'helpers/customHooks/useAmazonPayObjects';
-//import { useFormValidation } from 'helpers/customHooks/useFormValidation';
-import { AmazonPay } from 'helpers/forms/paymentMethods';
 import { setAmazonPayHasAccessToken } from 'helpers/redux/checkout/payment/amazonPay/actions';
 import {
 	useContributionsDispatch,
 	useContributionsSelector,
 } from 'helpers/redux/storeHooks';
-import {
-	trackComponentClick,
-	trackComponentLoad,
-} from 'helpers/tracking/behaviour';
-import { logException } from 'helpers/utilities/logger';
+import { trackComponentLoad } from 'helpers/tracking/behaviour';
 
 export function AmazonPaymentButton(): JSX.Element {
 	const dispatch = useContributionsDispatch();
-
-	const countryGroupId = useContributionsSelector(
-		(state) => state.common.internationalisation.countryGroupId,
-	);
-
-	const paymentMethod = useContributionsSelector(
-		(state) => state.page.checkoutForm.payment.paymentMethod.name,
-	);
-
-	const { loginObject, paymentsObject } = useAmazonPayObjects(
-		paymentMethod === AmazonPay,
-		countryGroupId,
-		false,
-	);
 
 	const hasAccessToken = useContributionsSelector(
 		(state) => state.page.checkoutForm.payment.amazonPay.hasAccessToken,
@@ -38,20 +16,7 @@ export function AmazonPaymentButton(): JSX.Element {
 	// const loginWithAmazonPay = useFormValidation(function login() {
 	function loginWithAmazonPay() {
 		console.log('amazonPaymentButton.loginWithAmazonPay');
-		trackComponentClick('amazon-pay-login-click');
-		const loginOptions = {
-			scope: 'profile postal_code payments:widget payments:shipping_address',
-			popup: true,
-		};
-		if (loginObject) {
-			loginObject.authorize(loginOptions, (response) => {
-				if (response.error) {
-					logException(`Error from Amazon login: ${response.error}`);
-				} else {
-					dispatch(setAmazonPayHasAccessToken());
-				}
-			});
-		}
+		dispatch(setAmazonPayHasAccessToken());
 	}
 	// });
 
@@ -59,13 +24,10 @@ export function AmazonPaymentButton(): JSX.Element {
 		console.log('amazonPaymentButton.payWithAmazonPay');
 	}
 
-	if (loginObject && paymentsObject) {
-		trackComponentLoad('amazon-pay-login-loaded');
-		return (
-			<DefaultPaymentButtonContainer
-				onClick={hasAccessToken ? payWithAmazonPay : loginWithAmazonPay}
-			/>
-		);
-	}
-	return <AnimatedDots appearance="dark" />;
+	trackComponentLoad('amazon-pay-login-loaded');
+	return (
+		<DefaultPaymentButtonContainer
+			onClick={hasAccessToken ? payWithAmazonPay : loginWithAmazonPay}
+		/>
+	);
 }
